Rename and move around pages and components
